refactor: move functions
